Fix typo in comment.

+++ sbbs3/mime.c
@@ -2,13 +2,13 @@
 
 /* Synchronet MIME functions */
 
-/* $Id: mime.c,v 1.8 2011/10/19 07:08:32 rswindell Exp $ */
+/* $Id: mime.c,v 1.7 2003/06/14 10:13:27 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2011 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2003 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -52,7 +52,7 @@ char* mimegetboundary()
     int		i, num;
     char*	boundaryString = (char*)malloc(SIZEOF_MIMEBOUNDARY + 1);
 
-    srand((unsigned int)time(NULL));
+    srand(time(NULL));
     if (boundaryString == NULL) 
         return NULL;
     for (i=0;i<SIZEOF_MIMEBOUNDARY;i++) {
